These changes are enough to get code to compile, and at least some of the unittests to pass, on FreeBSD 6.0, Solaris 10 x86, and OS X Leopard.  You have to use gmake instead of standard bsd make, though, because neither bsd nor solari make understand 'include */*.d' and 'CFLAGS += $(POPT_FLAGS).'  These may be fixable later.
Most changes fall into four categories:
1) #include differences
2) New errors due to some #ifdef paths being taken differently
3) Undefined functions (see, eg, the new HAVE_STRSEP)
4) Type differences (eg tv_usec is an int on os x, not a long int)

As one concrete example, snprintf.c is an empty file on linux, where
all the functionality is part of glibc.  But on Solaris, some of its
functionality is useful.  This turned up a bug where if you have
vsnprintf on your system but not vasnprintf, then dopr() was not being
used but was being defined, leading to an "unused static function"
warning in gcc.

As another, solaris would complain about "index" being used as a
variable, since it's also a function name.  The var was renamed idx.

Tested by compiling on FreeBSD, Solaris 10, and OS X leopard (x86).  I
also compiled and ran unittests under Linux Ubuntu, to make sure
this change didn't break anything there.

Reviewed by fergus@google.com

+++ src/compile.c
@@ -30,6 +30,7 @@
 #include <unistd.h>
 #include <fcntl.h>
 #include <errno.h>
+#include <sys/stat.h>
 
 #include <ctype.h>
 
@@ -680,7 +681,7 @@ int dcc_build_somewhere_timed(char *argv[],
 
         rs_log(RS_LOG_INFO|RS_LOG_NONAME,
                "elapsed compilation time %ld.%06lds",
-               delta.tv_sec, delta.tv_usec);
+               delta.tv_sec, (long) delta.tv_usec);
     }
 
     return ret;
